smart_inspection/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis
reportlab
Pillow
pytesseract
python-multipart
boto3
